added shared integration test setup code
